fix: minor fix

+++ Source/mechanics/mechanics.h
@@ -35,7 +35,7 @@ typedef struct {
 
 /* --- VARIABILI GLOBALI (EXTERN) --- */
 extern volatile uint8_t timer_tick;
-extern volatile int playing_field[HEIGHT][WIDTH];
+extern volatile uint16_t playing_field[HEIGHT][WIDTH];
 extern volatile int score;
 extern volatile int HighScore;
 extern volatile int game_started;
